requirements.txt
django
djangorestframework
djangorestframework-simplejwt
apscheduler
huggingface_hub
langchain
pinecone-client
python-dotenv
